+++ tabled_derive/src/attributes/mod.rs
@@ -1,0 +1,5 @@
+mod field_attr;
+mod type_attr;
+
+pub use field_attr::FieldAttributes;
+pub use type_attr::TypeAttributes;
