+++ tabled_derive/src/parse/mod.rs
@@ -1,2 +1,3 @@
 pub mod field_attr;
+pub mod func_arg;
 pub mod type_attr;
